Refactor a lot into useSarcophagusNegotiation; fixes; add and react to signaturesReady
