+++ src/features/embalm/stepContent/hooks/useCreateSarcohpagus.ts
@@ -61,8 +61,6 @@ export function useCreateSarcophagus() {
         shares: selectedArchaeologists.length,
         threshold: selectedArchaeologists.length,
       });
-      console.log('shards', shards);
-
 
       // Step 4: Encrypt each shard of the outer layer private key using each archaeologist's public
       // key
@@ -71,12 +69,12 @@ export function useCreateSarcophagus() {
       const encryptedShards = await encryptShards(archPublicKeys, shards);
 
       // Step 5: Upload the double encrypted payload to the arweave bundlr
-      const sarcophagusPayloadTxId = await uploadArweaveFile(encryptedOuterLayer);
+      const sarcophagusPayloadTxId = await uploadArweaveFile(encryptedOuterLayer); // TODO: change to use uploadFile for Bundlr
       // dispatch(setPayloadTxId(sarcophagusPayloadTxId));
 
       // Step 6: Upload the encrypted shards mapping to the arweave bundlr
       const mapping: Record<string, string> = encryptedShards.reduce((acc, shard) => ({ ...acc, [shard.publicKey]: shard.encryptedShard }), {});
-      const encryptedShardsTxId = await uploadArweaveFile(Buffer.from(JSON.stringify(mapping)));
+      const encryptedShardsTxId = await uploadArweaveFile(Buffer.from(JSON.stringify(mapping))); // TODO: change to use uploadFile for Bundlr
       // dispatch(setShardsTxId(encryptedShardsTxId));
 
       dispatch(setShardPayloadData(encryptedShards, sarcophagusPayloadTxId, encryptedShardsTxId));
